New Generation Settings
Systems generated can now have collisions disabled. This would not happen in the final build, but speeds up the generation time significantly. There is also a new generation method, which is controlled by the systemEquidistant boolean which allows the galaxy rings to be wider, and have more systems on them.