Writing data to CSV
Data within the game can now be written to a CSV, the first usage for this is galaxy generation optimisation analysis, where information about the galaxy is printed, along with the time it took to generate. This can then be used on other devices to see how long it takes to generate.