Minor Debug Changes
Small debug mode changes have been made, the number of systems to generate is now calculated and added to the CSV and also the ring intervals. The CSV has been renamed to GenerationData as it does not just give the time any more, and the ring interval multiplier can now be edited to produce more compact ring structured galaxies.